+++ .github/workflows/ci.yml
@@ -178,10 +178,10 @@ jobs:
       - uses: actions/checkout@v4
         with:
           submodules: recursive
-      - name: Set up Python 3.9
+      - name: Set up Python 3.12
         uses: actions/setup-python@v5
         with:
-          python-version: 3.9
+          python-version: 3.12
       - name: Install mypy
         run: pip install mypy
       - name: Run mypy
