+++ auto_tests/BUILD.bazel
@@ -4,6 +4,7 @@ cc_library(
     name = "check_compat",
     testonly = True,
     hdrs = ["check_compat.h"],
+    deps = ["//c-toxcore/toxcore:ccompat"],
 )
 
 cc_library(
@@ -14,11 +15,14 @@ cc_library(
     deps = [
         ":check_compat",
         "//c-toxcore/testing:misc_tools",
+        "//c-toxcore/toxcore:DHT",
         "//c-toxcore/toxcore:Messenger",
         "//c-toxcore/toxcore:mono_time",
+        "//c-toxcore/toxcore:net_crypto",
         "//c-toxcore/toxcore:tox",
         "//c-toxcore/toxcore:tox_dispatch",
         "//c-toxcore/toxcore:tox_events",
+        "//c-toxcore/toxcore:tox_log_level",
     ],
 )
 
@@ -72,6 +76,8 @@ extra_data = {
         "//c-toxcore/toxcore:onion",
         "//c-toxcore/toxcore:onion_announce",
         "//c-toxcore/toxcore:onion_client",
+        "//c-toxcore/toxcore:os_memory",
+        "//c-toxcore/toxcore:os_random",
         "//c-toxcore/toxcore:tox",
         "//c-toxcore/toxcore:tox_dispatch",
         "//c-toxcore/toxcore:tox_events",
