+++ other/BUILD.bazel
@@ -31,6 +31,8 @@ cc_binary(
         "//c-toxcore/toxcore:network",
         "//c-toxcore/toxcore:onion",
         "//c-toxcore/toxcore:onion_announce",
+        "//c-toxcore/toxcore:os_memory",
+        "//c-toxcore/toxcore:os_random",
         "//c-toxcore/toxcore:tox",
     ],
 )
