+++ other/bootstrap_daemon/BUILD.bazel
@@ -25,6 +25,7 @@ cc_binary(
         "//c-toxcore/toxcore:network",
         "//c-toxcore/toxcore:onion",
         "//c-toxcore/toxcore:onion_announce",
+        "//c-toxcore/toxcore:os_memory",
         "//c-toxcore/toxcore:os_random",
         "//c-toxcore/toxcore:tox",
         "@libconfig",
