+++ testing/support/BUILD.bazel
@@ -47,6 +47,7 @@ cc_library(
     visibility = ["//visibility:public"],
     deps = [
         "//c-toxcore/toxcore:mem",
+        "//c-toxcore/toxcore:net",
         "//c-toxcore/toxcore:network",
         "//c-toxcore/toxcore:rng",
         "//c-toxcore/toxcore:tox",
