+++ toxav/BUILD.bazel
@@ -17,7 +17,10 @@ cc_library(
     name = "ring_buffer",
     srcs = ["ring_buffer.c"],
     hdrs = ["ring_buffer.h"],
-    deps = ["//c-toxcore/toxcore:ccompat"],
+    deps = [
+        "//c-toxcore/toxcore:attributes",
+        "//c-toxcore/toxcore:ccompat",
+    ],
 )
 
 cc_test(
@@ -108,6 +111,7 @@ cc_test(
         ":bwcontroller",
         "//c-toxcore/toxcore:logger",
         "//c-toxcore/toxcore:mono_time",
+        "//c-toxcore/toxcore:network",
         "//c-toxcore/toxcore:os_memory",
         "@com_google_googletest//:gtest",
         "@com_google_googletest//:gtest_main",
@@ -156,6 +160,7 @@ cc_test(
         ":rtp",
         "//c-toxcore/toxcore:logger",
         "//c-toxcore/toxcore:mono_time",
+        "//c-toxcore/toxcore:network",
         "//c-toxcore/toxcore:os_memory",
         "@com_google_googletest//:gtest",
         "@com_google_googletest//:gtest_main",
@@ -187,6 +192,7 @@ cc_test(
         ":video",
         "//c-toxcore/toxcore:logger",
         "//c-toxcore/toxcore:mono_time",
+        "//c-toxcore/toxcore:network",
         "//c-toxcore/toxcore:os_memory",
         "@com_google_googletest//:gtest",
         "@com_google_googletest//:gtest_main",
@@ -218,6 +224,7 @@ cc_binary(
         ":rtp",
         "//c-toxcore/toxcore:logger",
         "//c-toxcore/toxcore:mono_time",
+        "//c-toxcore/toxcore:network",
         "//c-toxcore/toxcore:os_memory",
         "@benchmark",
     ],
@@ -255,6 +262,7 @@ cc_test(
     srcs = ["msi_test.cc"],
     deps = [
         ":msi",
+        "//c-toxcore/toxcore:logger",
         "//c-toxcore/toxcore:os_memory",
         "@com_google_googletest//:gtest",
         "@com_google_googletest//:gtest_main",
