+++ toxcore/BUILD.bazel
@@ -431,6 +431,30 @@ cc_library(
 )
 
 cc_library(
+    name = "net",
+    srcs = ["net.c"],
+    hdrs = ["net.h"],
+    visibility = ["//c-toxcore:__subpackages__"],
+    deps = [
+        ":attributes",
+        ":mem",
+    ],
+)
+
+cc_library(
+    name = "os_network",
+    srcs = ["os_network.c"],
+    hdrs = ["os_network.h"],
+    visibility = ["//c-toxcore:__subpackages__"],
+    deps = [
+        ":attributes",
+        ":ccompat",
+        ":mem",
+        ":net",
+    ],
+)
+
+cc_library(
     name = "network",
     srcs = [
         "net_log.c",
@@ -444,6 +468,7 @@ cc_library(
         "//c-toxcore/auto_tests:__subpackages__",
         "//c-toxcore/other:__pkg__",
         "//c-toxcore/other/bootstrap_daemon:__pkg__",
+        "//c-toxcore/testing/bench:__pkg__",
         "//c-toxcore/testing/fuzzing:__pkg__",
         "//c-toxcore/testing/support:__pkg__",
         "//c-toxcore/toxav:__pkg__",
@@ -456,7 +481,9 @@ cc_library(
         ":logger",
         ":mem",
         ":mono_time",
+        ":net",
         ":net_profile",
+        ":os_network",
         ":util",
         "@libsodium",
         "@psocket",
@@ -472,6 +499,7 @@ cc_library(
     deps = [
         ":crypto_core",
         ":mem",
+        ":net",
         ":network",
         ":rng",
         ":test_util",
@@ -546,6 +574,7 @@ cc_library(
         ":ccompat",
         ":crypto_core",
         ":mem",
+        ":net",
         ":network",
         ":util",
         "@psocket",
@@ -563,6 +592,7 @@ cc_library(
         "ping.h",
     ],
     visibility = [
+        "//c-toxcore/auto_tests:__pkg__",
         "//c-toxcore/other:__pkg__",
         "//c-toxcore/other/bootstrap_daemon:__pkg__",
         "//c-toxcore/testing:__pkg__",
@@ -576,6 +606,7 @@ cc_library(
         ":logger",
         ":mem",
         ":mono_time",
+        ":net",
         ":network",
         ":ping_array",
         ":rng",
@@ -597,6 +628,7 @@ cc_library(
         ":crypto_core_test_util",
         ":logger",
         ":mono_time",
+        ":net",
         ":net_crypto",
         ":network",
         ":network_test_util",
@@ -656,6 +688,7 @@ cc_library(
         ":logger",
         ":mem",
         ":mono_time",
+        ":net",
         ":network",
         ":rng",
         ":shared_key_cache",
@@ -680,6 +713,7 @@ cc_library(
         ":logger",
         ":mem",
         ":mono_time",
+        ":net",
         ":network",
         ":rng",
         ":timed_auth",
@@ -734,6 +768,7 @@ cc_library(
         ":crypto_core",
         ":logger",
         ":mem",
+        ":net",
         ":net_profile",
         ":network",
         ":rng",
@@ -763,6 +798,7 @@ cc_library(
         ":logger",
         ":mem",
         ":mono_time",
+        ":net",
         ":net_profile",
         ":network",
         ":onion",
@@ -786,6 +822,7 @@ cc_library(
         ":logger",
         ":mem",
         ":mono_time",
+        ":net",
         ":net_profile",
         ":network",
         ":rng",
@@ -810,6 +847,7 @@ cc_library(
         ":logger",
         ":mem",
         ":mono_time",
+        ":net",
         ":net_profile",
         ":network",
         ":onion",
@@ -868,6 +906,7 @@ cc_library(
         ":logger",
         ":mem",
         ":mono_time",
+        ":net",
         ":net_profile",
         ":network",
         ":rng",
@@ -948,6 +987,7 @@ cc_library(
         ":logger",
         ":mem",
         ":mono_time",
+        ":net",
         ":network",
         ":onion",
         ":rng",
@@ -976,6 +1016,7 @@ cc_library(
         ":logger",
         ":mem",
         ":mono_time",
+        ":net",
         ":network",
         ":util",
     ],
@@ -1054,6 +1095,7 @@ cc_library(
         ":logger",
         ":mem",
         ":mono_time",
+        ":net",
         ":net_crypto",
         ":network",
         ":onion",
@@ -1097,6 +1139,7 @@ cc_fuzz_test(
         ":DHT",
         ":net_crypto",
         ":net_profile",
+        ":network",
         ":onion_client",
         "//c-toxcore/testing/support",
     ],
@@ -1117,6 +1160,7 @@ cc_library(
         ":logger",
         ":mem",
         ":mono_time",
+        ":net",
         ":net_crypto",
         ":network",
         ":onion",
@@ -1241,6 +1285,7 @@ cc_library(
         ":logger",
         ":mem",
         ":mono_time",
+        ":net",
         ":net_crypto",
         ":net_profile",
         ":network",
@@ -1324,11 +1369,13 @@ cc_library(
         ":logger",
         ":mem",
         ":mono_time",
+        ":net",
         ":net_crypto",
         ":net_profile",
         ":network",
         ":onion_client",
         ":os_memory",
+        ":os_network",
         ":os_random",
         ":state",
         ":tox_attributes",
