+++ toxcore/Messenger.h
@@ -27,6 +27,7 @@
 #include "logger.h"
 #include "mem.h"
 #include "mono_time.h"
+#include "net.h"
 #include "net_crypto.h"
 #include "net_profile.h"
 #include "network.h"
