+++ toxcore/forwarding.h
@@ -14,6 +14,7 @@
 #include "logger.h"
 #include "mem.h"
 #include "mono_time.h"
+#include "net.h"
 #include "network.h"
 #include "rng.h"
