+++ toxcore/friend_connection.h
@@ -17,6 +17,7 @@
 #include "logger.h"
 #include "mem.h"
 #include "mono_time.h"
+#include "net.h"
 #include "net_crypto.h"
 #include "network.h"
 #include "onion_client.h"
