+++ toxcore/group_announce.h
@@ -18,6 +18,7 @@
 #include "logger.h"
 #include "mem.h"
 #include "mono_time.h"
+#include "net.h"
 #include "network.h"
 
 #ifdef __cplusplus
