+++ toxcore/group_connection.h
@@ -21,6 +21,7 @@
 #include "logger.h"
 #include "mem.h"
 #include "mono_time.h"
+#include "net.h"
 #include "network.h"
 
 /* Max number of TCP relays we share with a peer on handshake */
