+++ toxcore/network_test_util.hh
@@ -6,6 +6,7 @@
 
 #include "crypto_core.h"
 #include "mem.h"
+#include "net.h"
 #include "network.h"
 #include "rng.h"
 #include "test_util.hh"
