+++ toxcore/onion.h
@@ -17,6 +17,7 @@
 #include "logger.h"
 #include "mem.h"
 #include "mono_time.h"
+#include "net.h"
 #include "network.h"
 #include "rng.h"
 #include "shared_key_cache.h"
