+++ toxcore/ping.h
@@ -17,6 +17,7 @@
 #include "crypto_core.h"
 #include "mem.h"
 #include "mono_time.h"
+#include "net.h"
 #include "network.h"
 
 typedef struct Ping Ping;
